refactor: change export functions of updaters for user account and settings
