+++ src/store/account/index.ts
@@ -6,3 +6,4 @@
 
 export * from './hooks';
 export * from './store';
+export * from './updaters';
